Resolves Issue #40
